+++ templates/modelRelation1To1/Model/{{ entity2Name|pascal }}.php
@@ -14,6 +14,8 @@ use {{ vendorName|pascal }}\{{ moduleName|pascal }}\Model\ResourceModel\{{ entit
 class {{ entity2Name|pascal }} extends AbstractModel implements {{ entity2Name|pascal }}Interface
 {
 
+    protected $_eventPrefix = '{{ vendorName|snake }}_{{ moduleName|snake }}_{{ entityName|snake }}';
+
     /**
      * @return int|null
      */